+++ NEWS.md
@@ -1,7 +1,8 @@
 # medfate 5.1.0
 * Compilation patch for C++ 20
-+ Mistletoe effects on water balance
-+ Recovery from defoliation depends on relative growth rate of buds 
+* Mistletoe effects on water balance
+* Recovery from defoliation depends on relative growth rate of buds 
+* Water potential at turgor loss point as estimation of Gsw_P50_Baldocchi
 
 # medfate 5.0.0
 * Partial rhizosphere overlap set to default